chore: clean up dependencies to reduce binary size
